requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
async-timeout==4.0.3
attrs==26.1.0
backports.asyncio.runner==1.2.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.25.2
Flask==3.1.1
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.4.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-huggingface==1.2.1
langchain-ollama==1.1.0
langchain-openai==1.1.12
langchain-pinecone==0.2.13
langchain-tests==0.3.22
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
ollama==0.6.1
openai==2.31.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
parso==0.8.5
pillow==12.2.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
pip==26.0.1
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
pydantic-settings==2.13.1
Pygments==2.19.2
pypdf==5.6.1
pytest==8.4.2
pytest-asyncio==1.3.0
pytest-benchmark==5.2.3
pytest-codspeed==4.3.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.10.2
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
stack_data==0.6.3
sympy==1.14.0
syrupy==4.9.1
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.11.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid_utils==0.14.1
vcrpy==7.0.0
wcwidth==0.2.14
Werkzeug==3.1.8
wheel==0.46.3
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
